refactor: add Login Container to separate logic
