+++ package.json
@@ -30,7 +30,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-router-hash-link": "^1.2.1",
-    "react-scripts": "3.4.1",
+    "react-scripts": "^3.4.3",
     "react-toastify": "^5.5.0",
     "typescript": "^3.9.6"
   },
