+++ src/firebase.ts
@@ -1,6 +1,4 @@
-import firebase from 'firebase/app'
-import 'firebase/auth'
-import 'firebase/firestore'
+import * as firebase from 'firebase';
 
 let config = null
