+++ src/pages/Login/Login.tsx
@@ -1,68 +1,28 @@
-import React, { useState, useContext, useReducer } from 'react'
+import React from 'react'
 import { motion } from 'framer-motion'
-import { withRouter } from 'react-router-dom'
-import { UserContext } from '../../components/UserContext'
-import { apiReducer, apiInitialState } from '../../reducers/apiReducer'
-import { Navbar } from '../../ui/Navbar/Navbar'
-import { ADMIN } from '../../constants/routes'
-import { navigateTo } from '../../utils/navigate'
-import { login } from '../../utils/login'
-import { notifyError } from '../../utils/notification'
-import { DB_ERROR_MSG } from '../../constants/toastMessages'
 import { pageTransitions } from '../../constants/config'
 import bookTableImg from '../../assets/images/brooke-lark-book-table.jpg'
+import { Navbar } from '../../ui/Navbar/Navbar'
 import { Spinner } from '../../ui/Spinner/Spinner'
 import { Button } from '../../ui/Button/Button'
 import { Label } from '../../ui/Label/Label'
 import { Input } from '../../ui/Input/Input'
 
-type Error = {
-  message: string
+type Props = {
+  onSubmit: (e: React.FormEvent<HTMLFormElement>) => Promise<void>
+  handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void
+  state: {
+    data: any
+    isLoading: boolean
+    error: null
+  }
+  error: {
+    message: string
+  }
 }
 
-const Login: React.FC<any> = ({ history }) => {
-  const [state, dispatch] = useReducer(apiReducer, apiInitialState)
+export const Login: React.FC<Props> = ({ onSubmit, state, handleChange, error }) => {
   const links = [{ name: 'Menu', link: 'menu' }, { name: 'Book Table', link: 'book-table' }]
-  const [error, setError] = useState<any>({
-    error: {
-      message: ''
-    }
-  })
-
-  const [form, setInputs] = useState({
-    email: '',
-    password: ''
-  })
-
-  const { setUser } = useContext(UserContext)
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    dispatch({ type: 'FETCHING' })
-    e.preventDefault()
-    try {
-      const user = await login(form.email, form.password)
-      if (setUser) {
-        setUser(user)
-        dispatch({ type: 'SUCCESS' })
-        navigateTo(history, ADMIN)
-      }
-    } catch (error) {
-      dispatch({ type: 'ERROR' })
-      handleError(error)
-      notifyError(DB_ERROR_MSG)
-    }
-  }
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setInputs({
-      ...form,
-      [e.target.name]: e.target.value
-    })
-  }
-
-  const handleError = (error: any) => {
-    setError(error)
-  }
 
   if (state.isLoading) {
     return <Spinner />
@@ -74,15 +34,15 @@ const Login: React.FC<any> = ({ history }) => {
       <motion.div initial="exit" animate="enter" exit="exit" variants={pageTransitions} className="container row">
         <div className="section section__col login">
           <h1 className="heading">Login</h1>
-          <form method="POST" className="login__form" onSubmit={handleSubmit}>
+          <form method="POST" className="login__form" onSubmit={onSubmit}>
             <Label htmlFor="email" className="label login__label">
               Email
             </Label>
-            <Input type="email" name="email" required onChange={handleInputChange} />
+            <Input type="email" name="email" required onChange={handleChange} />
             <Label htmlFor="email" className="label login__label">
               Password
             </Label>
-            <Input type="password" name="password" required onChange={handleInputChange} />
+            <Input type="password" name="password" required onChange={handleChange} />
             <p className="login__error">{error ? error.message : null}</p>
             <Button data-testid="login-submit" type="submit" className="btn--dark">
               Login
@@ -98,5 +58,3 @@ const Login: React.FC<any> = ({ history }) => {
     </>
   )
 }
-
-export default withRouter(Login)
